Rename chatbotCli to chatbot
